02-challenge/01_Implementation/01_Implementation_Code/Dataset2_Implementation/requirements.txt
langchain==1.2.15
langchain-core==1.2.28
langchain-openai==1.1.12
langgraph==1.1.6
langgraph-prebuilt==1.0.9
langfuse==4.1.0
python-dotenv==1.2.2
ulid-py==1.1.0
